👽️ `stats.{siegel,theil}slopes`: move from `_stats_mstats_common` to `_correlation`
